+++ kernel/lib/libc/mem.h
@@ -4,7 +4,7 @@
 
 EXTERN_C_START
 
-int memcmp(const void* buf1, const void* buf2, unsigned long long n);
+int memcmp(const void* buf1, const void* buf2, size_t n);
 void* memcpy(void* dst, const void* src, size_t n);
 void* memset(void* dst, u32 val, size_t n);
